float->double

+++ src/solver/simulation/sim_structure_probleme_economique.h
@@ -501,8 +501,8 @@ struct AdequacyPatchParameters
     bool SaveIntermediateResults;
     bool IncludeHurdleCostCsr;
     AdqPatchPTO PriceTakingOrder;
-    float ThresholdInitiateCurtailmentSharingRule;
-    float ThresholdDisplayLocalMatchingRuleViolations;
+    double ThresholdInitiateCurtailmentSharingRule;
+    double ThresholdDisplayLocalMatchingRuleViolations;
 };
 
 struct PROBLEME_HEBDO
